account timeline cleanup
